requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.0
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.47.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.29.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.2.1
coloredlogs==15.0.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
email_validator==2.2.0
fastapi==0.115.13
filelock==3.18.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.8.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.25.1
google-api-python-client==2.150.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kubernetes==33.1.0
limits==5.4.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.5.0
neo4j==5.28.1
numpy==1.26.4
oauthlib==3.3.0
onnxruntime==1.22.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
ordered-set==4.1.0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pgvector==0.3.5
posthog==5.1.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.9
pulsar-client==3.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.1
PyJWT==2.8.0
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
sympy==1.14.0
tenacity==9.0.0
tokenizers==0.21.1
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.0
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
